io: load NIfTI at float32 instead of float64
nibabel's get_fdata() default inflates an int16-on-disk NIfTI 4x (the
61 MB test scan became a 245 MB float64 array). Request float32 instead:
every other loader (Philips REC, BART, text) already delivers
float32/complex64, all three FFT backends preserve float32 -> complex64,
and the LOD pyramid reduces to float32/complex64 regardless of input, so
the narrow dtype now cascades through the whole pipeline. Measured on the
real scan shape: source memory 245.7 -> 122.8 MB and scipy centered-FFT
axis-0 113 -> 65 ms.

+++ arrayscope/io/file_interpreters.py
@@ -602,7 +602,10 @@ class NiftiLoader:
             ) from err
 
         image = nib.load(self.file_path)
-        data = image.get_fdata()
+        # float32 keeps the whole downstream pipeline narrow (FFT -> complex64,
+        # LOD pyramid float32); NIfTI voxels are almost always int16/float32 on
+        # disk, so float64 here would only inflate memory 2-4x.
+        data = image.get_fdata(dtype=np.float32)
         data = remove_trailing_singletons(data)
         self.axes = _nifti_axes(image.header, data.shape)
         self.metadata.update(
